Refactor Vulkan Backend Operations and Enhance Gradient Compression
- Updated Vulkan backend memory operations to use double precision for fill and full operations, ensuring better handling of Float64 subnormals.
- Modified the dispatch functions in Vulkan kernel registry to accommodate the new double precision parameters.
- Adjusted the dispatchSearchSorted function to return Int64 by default, aligning with CPU/CUDA conventions.
- Enhanced the TopKCompressor to maintain the gradient buffer's data pointer across iterations, preventing residual cache misses.
- Improved LayerNorm and pooling layers to ensure contiguous input tensors, optimizing performance across backends.
- Updated tests to reflect changes in data types and ensure compatibility with new backend operations.
- Added checks for backend availability in unit tests to prevent unnecessary failures when devices are not present.

+++ src/backends/cuda/cuda_backend.cpp
@@ -165,7 +165,7 @@ namespace cuda {
     auto zeros_kernel(const std::vector<int64_t>& shape, DType dtype, Device device, cudaStream_t stream) -> Tensor;
     auto ones_kernel(const std::vector<int64_t>& shape, DType dtype, Device device, cudaStream_t stream) -> Tensor;
     auto full_kernel(const std::vector<int64_t>& shape, float value, DType dtype, Device device, cudaStream_t stream) -> Tensor;
-    auto fill_kernel(const Tensor& tensor, float value, cudaStream_t stream) -> Tensor;
+    auto fill_kernel(const Tensor& tensor, double value, cudaStream_t stream) -> Tensor;
 
     // Random operations
     auto rand_kernel(const std::vector<int64_t>& shape, DType dtype, Device device, cudaStream_t stream) -> Tensor;
