+++ src/backends/cuda/cuda_kernel_registry.cpp
@@ -524,8 +524,8 @@ namespace cuda {
 #endif
 
     // Fill operations
-    auto fill_kernel(const Tensor& tensor, float value, cudaStream_t stream) -> Tensor;
-    auto strided_fill_kernel(Tensor& self, float value, cudaStream_t stream) -> void;
+    auto fill_kernel(const Tensor& tensor, double value, cudaStream_t stream) -> Tensor;
+    auto strided_fill_kernel(Tensor& self, double value, cudaStream_t stream) -> void;
 
     // Runtime cuDNN availability check
     bool is_cudnn_available() noexcept;
@@ -1348,11 +1348,11 @@ void register_cuda_kernels(BackendDispatchTable& table) {
         return cuda::clone_kernel(inputs[0], get_cuda_stream(attrs));
     });
     table.register_single_output_kernel(OpId::Fill, [](std::span<const Tensor> inputs, const OpAttributes& attrs) -> Tensor {
-        float value = static_cast<float>(attrs.get_float(AttrKey::Value, 0.0));
+        double value = attrs.get_float(AttrKey::Value, 0.0);
         return cuda::fill_kernel(inputs[0], value, get_cuda_stream(attrs));
     });
     table.register_inplace_kernel(OpId::StridedFill, [](Tensor& self, std::span<const Tensor>, const OpAttributes& attrs) -> Tensor& {
-        float value = static_cast<float>(attrs.get_float(AttrKey::Value, 0.0));
+        double value = attrs.get_float(AttrKey::Value, 0.0);
         cuda::strided_fill_kernel(self, value, get_cuda_stream(attrs));
         return self;
     });
@@ -2940,7 +2940,9 @@ void register_cuda_kernels(BackendDispatchTable& table) {
     });
     table.register_kernel(OpId::Full, [](std::span<const Tensor> inputs, const OpAttributes& attrs) {
         auto shape = attrs.get_int_list(AttrKey::Shape);
-        float value = static_cast<float>(attrs.get_float(AttrKey::Value, 0.0));
+        // Read as double (the attribute is stored as double) so Float64
+        // subnormals survive the trip through OpAttributes.
+        double value = attrs.get_float(AttrKey::Value, 0.0);
         DType dtype = dtype_from_string(attrs.get_string(AttrKey::Dtype, "float32"));
         int device_idx = static_cast<int>(attrs.get_int(AttrKey::Device, 0));
         Device device = Device::cuda(device_idx);
