+++ src/backends/rocm/rocm_backend.hpp
@@ -91,8 +91,8 @@ namespace rocm {
     // Fill operations
     auto zeros_kernel(const std::vector<int64_t>& shape, DType dtype, Device device, hipStream_t stream) -> Tensor;
     auto ones_kernel(const std::vector<int64_t>& shape, DType dtype, Device device, hipStream_t stream) -> Tensor;
-    auto full_kernel(const std::vector<int64_t>& shape, float value, DType dtype, Device device, hipStream_t stream) -> Tensor;
-    auto fill_kernel(const Tensor& tensor, float value, hipStream_t stream) -> Tensor;
+    auto full_kernel(const std::vector<int64_t>& shape, double value, DType dtype, Device device, hipStream_t stream) -> Tensor;
+    auto fill_kernel(const Tensor& tensor, double value, hipStream_t stream) -> Tensor;
 
     // Random operations
     auto rand_kernel(const std::vector<int64_t>& shape, DType dtype, Device device, hipStream_t stream) -> Tensor;
